Separate out Edge image terms test to a new class
